+++ CHANGES.md
@@ -3,6 +3,7 @@
 ### Version 0.2.2
 
 - Added SN76489 stereo separation to mixer menu
+- Added `NE0` / `NE1` noise reset enable effect
 - SN76489 VGM logger now eliminates extra register writes that have no side effects
 - Extended VGM header size so that it will not be misinterpreted by certain players
 - Fixed text export and import (for SN7T only)